Fix flippy bridge logic.
'Connection's don't make sense as a key since they only model half
of the connection, and even with both sides and the type of bridge
you can have multiple bridges between the same zones. We solve this
the way you usually do: with a multiset. This is tougher to abstract
into the BridgesState object though, so we drop it entirely.